rename services to price requests and assertions
