+++ libs/src/oracle2/services/assertions/gql/factory.ts
