+++ libs/src/oracle2/services/assertions/gql/index.ts
