+++ libs/src/oracle2/services/assertions/gql/utils.ts
