+++ libs/src/oracle2/services/assertions/index.ts
