+++ libs/src/oracle2/services/index.ts
@@ -1,4 +1,4 @@
-export * as oracleV1Gql from "./oracleV1/gql";
-export * as oracleV3Gql from "./oracleV3/gql";
+export * as oracleV3Gql from "./assertions/gql";
 export * as oracles from "./oracles";
+export * as oracleV1Gql from "./priceRequests/gql";
 export * as tokens from "./tokens";
