+++ libs/src/oracle2/services/priceRequests/gql/factory.ts
