+++ libs/src/oracle2/services/priceRequests/gql/index.ts
