+++ libs/src/oracle2/services/priceRequests/gql/queries.ts
