+++ libs/src/oracle2/services/priceRequests/gql/utils.ts
